requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==1.4.1
boto==2.27.0
celery==3.1.18
dj-database-url==0.3.0
Django==1.7.10
django-celery==3.1.16
django-picklefield==0.3.0
django-secure==1.0.1
django-ses==0.6.0
django-sslify==0.2.7
django-webtest==1.7.8
gunicorn==19.3.0
html5lib==0.999
httplib2==0.10.3
kombu==3.0.30
oauth2client==1.2
psycopg2==2.7.1
pyparsing==1.5.7
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
requests==2.7.0
rollbar==0.12.1
six==1.10.0
unicodecsv==0.9.0
validate-email==1.2
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.18
pycrypto==2.6.1
